pyproject.toml
[project]
name = "ai-agent-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.3",
    "mcp[cli]>=1.17.0",
    "openai>=2.0.1",
    "openai-agents>=0.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
